Signal-centric ECU redesign: multi-mode PID stores ($1A/$22/$2D) + OBD-II Mode $01
Restructures the ECU model around shared signal/identity/state layers with
per-mode PID projections, replacing the single flat Pid store.

- EcuNode gains separate per-mode stores (mode1APids/mode22Pids/mode2DPids)
  keyed by Address; AddPid/RemovePid/RelocatePidMode route by PidMode.
- OBD-II Mode $01 (J1979): Service01Handler + J1979 catalogue projection over
  the signal layer + per-ECU Mode1Supported set.
- Editor multi-mode PID grid: per-mode sections ($1A/$22/$2D) replace the Mode
  column; rows are stamped by the section that adds them.
- EcuMode22Seeder / EcuIdentitySeeder seed sensible default rows on new ECUs.
- CSV waveform replay, Ford-capture persona + PersonaRegistry, bin family
  classifier + BinEcuFactory, UniversalPatcher XML autodetect, PowerPc sniffer.

Fix: editing a PID's Address in the editor now re-keys its per-mode store via
EcuNode.RekeyPidAddress. Previously the dict entry stayed under the old key,
so GetPid/GetPidByWireId missed the new address and $2D/$22 reads returned NRC
$31 RequestOutOfRange even though the saved config looked correct. Covered by
Tests.Unit/Ecu/RekeyPidAddressTests.cs.

+++ Common/Protocol/Gmw3110Annotator.cs
@@ -43,7 +43,7 @@ public static class Gmw3110Annotator
         if (len < 1 || payload.Length < offset + 1 + len) return false;
         byte sid = payload[offset + 1];
         return sid == Service.TesterPresent
-            || sid == (byte)(Service.TesterPresent + 0x40);
+            || sid == Service.TesterPresent + 0x40;
     }
 
     /// <summary>
